sync config
removing duplicated configuration

the config file tree was duplicated, but the got out of sync.  This patch synchronizes the key, prior to removing the duplicated code

the java package: com.netscape.certsrv.common contains configuration information used by both the console and the server.  It was copied in the source tree, and was getting duplicated.  This removes the console version of the source, and instead copies the Java files prior to compilation