eclipse project files
This contains only the files necessary to get the project to import into Eclipse,  plus a very small .gitignore file for people that want to use git svn.  This patch should have no effects on the way that code currently runs.

The difference between this patch and the previous submission is that this one does not attempt to deal with the problems due to the code duplication between the console and common code bases:  it does not include the console in the classpath.  People that want to use eclipse to work with the console will have to perform additional modifications.  Also, it does not include the  directory base/common/test that has mock objects for abstract base classes that have not been completely implemented.

This version has removed all .class file references from the .gitignore as well.