Dead code removal
This patch removes all of the locations that Eclipse identified as Dead code.
Only the Eclipse automated cleanups were performed,  which means that some of the locations which were in *if* blocks still have the corresponding brackets around them.
These ensure that the original variable semantics weren't changed,  but are safe to remove in the future.