https://fedorahosted.org/pki/attachment/ticket/237/ ECC support for tpsclient two parameters are added in the config file : slotnamefile - this points to a file that contains the slot name. The reason why the slot name goes into another file instead of the conf file is because the existing conf file uses spaces for deliminator and I don't want to change the usage (and QE's tests) by changing it to something else. tokpasswod is the password to the token
This ECC feature requires the up-coming ECC supporting TPS to trigger. Testing of the ECC specific areas requires the availability of such ECC TPS. Without it, one can test with the old RSA method to see if it's broken. In addition to testing for ECC and RSA functionality in regards to tpsclient, I also tested it against the up coming DRM/TPS for server-side key generation and archival ECC support.