Resolves: rhbz #1454414
Bugzilla Bug #1454414 - ECDSA Certificates Generated by Certificate System 8.1
fail NIST validation test with parameter field

Fix generated by cfu in RHEL 7; backported to RHCS 8.1.8.

Ticket #1741 ECDSA certs Alg IDs contian parameter field

Per rfc5758, When the ecdsa-with-SHA224, ecdsa-with-SHA256, ecdsa-with-SHA384,
or ecdsa-with-SHA512 algorithm identifier appears in the algorithm field as an
AlgorithmIdentifier, the encoding MUST omit the parameters field.
Note: Since we do not support DSA, this patch does not attempt to address them.

Also, while we do not claim to support sha224, the patch adds enough code to
process the OID just for completeness.  However, it does not attempt to offer
it as part of the signing algorithms.